scraper/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.4
pyyaml==6.0.1
aiofiles==23.2.1
